+++ src/main/java/com/revature/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package com.revature.repository;
+package com.revature.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
