+++ src/main/java/com/revature/services/UserService.java
@@ -1,4 +1,4 @@
-package com.revature.service;
+package com.revature.services;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.revature.models.AppUser;
-import com.revature.repository.UserRepository;
+import com.revature.repositories.UserRepository;
 
 /**
  * This class contains methods that should be accessed by the controller to find and edit users.
