+++ src/test/java/com/revature/controllers/AuthControllerTest.java
@@ -1,4 +1,4 @@
-package com.revature.tests;
+package com.revature.controllers;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -8,8 +8,6 @@ import static org.mockito.Mockito.when;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
 
 import org.junit.Before;
 import org.junit.Ignore;
@@ -23,24 +21,14 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.junit.rules.ExpectedException;
 import org.springframework.security.core.Authentication;
 
-import com.revature.controller.AuthController;
+import com.revature.controllers.AuthController;
 import com.revature.exceptions.UserCreationException;
 import com.revature.exceptions.UserNotFoundException;
 import com.revature.models.AppUser;
 
 import com.revature.models.UserPrincipal;
+import com.revature.services.UserService;
 import com.revature.models.UserErrorResponse;
-import com.revature.service.UserService;
-import com.revature.security.CustomAuthenticationFilter;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.web.servlet.MockMvc;
-
-
-
 /**
  * This class will be used to test methods in the AuthController. Testing will
  * be done using Junit to test different scenarios.
@@ -48,20 +36,9 @@ import org.springframework.test.web.servlet.MockMvc;
  * @RunWith(): Specifies that we will be using the MockitoJUnit runner class.
  */
 @RunWith(MockitoJUnitRunner.class)
-
-@WebMvcTest(AuthController.class)
 public class AuthControllerTest {
 
-	private MockMvc mvc;
-	
-	
 
-	
-	@InjectMocks
-	private CustomAuthenticationFilter filter;
-	
-	@Mock
-	HttpServletRequest request;
 	@Mock
 	private UserService userService;
 	@Mock
@@ -93,7 +70,6 @@ public class AuthControllerTest {
 	}
 	
 	/**
-
 	 * This test case verifies proper functionality of the AuthController.getAllUsers() method.
 	 * A non-null ArrayList of AppUser objects with a size of one is expected to be returned.
 	 * 
@@ -477,43 +453,4 @@ public class AuthControllerTest {
 		assertEquals("The UserErrorResponse.message is expected to match the mocked message", "Mocked user creation exception", testResult.getMessage());
 	}
 	
-
-
-
-
-/**
- * This method tests integration of the controller with MockMVC
- * 
- * @author Jaitee Pitts
- * @author Ankit Patel
- *
- */
-
-
-	
-
-	
-	
-	
-	/**
-	 * Tests that delete user cannot be called without a user role
-	 * @throws Exception
-	 */
-	@Test
-	public void deleteUserUnauthorized() throws Exception{
-		this.mvc.perform(delete("/id/0")).andExpect(status().isUnauthorized());
-		
-	}
-	/**
-	 * Tests that delete user cannot be called without a Zuul header
-	 * @throws Exception
-	 * 
-	 */
-	@Test
-	@WithMockUser(roles= {"USER"})
-	public void SubVersionAttemptTest() throws Exception{
-		this.mvc.perform(delete("/id/0")).andExpect(status().isUnauthorized());
-	}
-	
-
 }
