+++ src/test/java/com/revature/controllers/ControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.revature.tests;
+package com.revature.controllers;
 
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -18,10 +18,10 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
-import com.revature.controller.AuthController;
+import com.revature.controllers.AuthController;
 import com.revature.models.AppUser;
 import com.revature.security.CustomAuthenticationFilter;
-import com.revature.service.UserService;
+import com.revature.services.UserService;
 
 
 /**
