+++ src/test/java/com/revature/controllers/EmailUsernameTest.java
@@ -1,4 +1,4 @@
-package com.revature.tests;
+package com.revature.controllers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
@@ -15,9 +15,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.revature.controller.AuthController;
+import com.revature.controllers.AuthController;
 import com.revature.models.AppUser;
-import com.revature.service.UserService;
+import com.revature.services.UserService;
 
 /**
  * Class containing MockMVC tests for the email and username validation methods.
